+++ addOns/authhelper/src/test/java/org/zaproxy/addon/authhelper/SessionDetectionScanRuleUnitTest.java
@@ -209,7 +209,7 @@ class SessionDetectionScanRuleUnitTest extends PassiveScannerTest<SessionDetecti
         assertThat(alertsRaised.size(), is(equalTo(1)));
         assertThat(alertsRaised.get(0).getUri(), is(equalTo("https://example0/login.jsp")));
         assertThat(alertsRaised.get(0).getMethod(), is(equalTo("GET")));
-        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("sanitizedtoken0")));
+        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("JSESSIONID")));
         assertThat(alertsRaised.get(0).getOtherInfo(), is(equalTo("\ncookie:JSESSIONID")));
         assertThat(alertsRaised.get(0).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
     }
@@ -238,17 +238,17 @@ class SessionDetectionScanRuleUnitTest extends PassiveScannerTest<SessionDetecti
         assertThat(alertsRaised.size(), is(equalTo(3)));
         assertThat(alertsRaised.get(0).getUri(), is(equalTo("https://example0/login")));
         assertThat(alertsRaised.get(0).getMethod(), is(equalTo("GET")));
-        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("sanitizedtoken17")));
+        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("session")));
         assertThat(alertsRaised.get(0).getOtherInfo(), is(equalTo("\ncookie:session")));
         assertThat(alertsRaised.get(0).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
         assertThat(alertsRaised.get(1).getUri(), is(equalTo("https://example0/dashboard")));
         assertThat(alertsRaised.get(1).getMethod(), is(equalTo("GET")));
-        assertThat(alertsRaised.get(1).getEvidence(), is(equalTo("sanitizedtoken25")));
+        assertThat(alertsRaised.get(1).getEvidence(), is(equalTo("session")));
         assertThat(alertsRaised.get(1).getOtherInfo(), is(equalTo("\ncookie:session")));
         assertThat(alertsRaised.get(1).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
         assertThat(alertsRaised.get(2).getUri(), is(equalTo("https://example0/login")));
         assertThat(alertsRaised.get(2).getMethod(), is(equalTo("GET")));
-        assertThat(alertsRaised.get(2).getEvidence(), is(equalTo("sanitizedtoken25")));
+        assertThat(alertsRaised.get(2).getEvidence(), is(equalTo("session")));
         assertThat(alertsRaised.get(2).getOtherInfo(), is(equalTo("\ncookie:session")));
         assertThat(alertsRaised.get(2).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
     }
@@ -276,7 +276,7 @@ class SessionDetectionScanRuleUnitTest extends PassiveScannerTest<SessionDetecti
         // Then
         assertThat(alertsRaised.size(), is(equalTo(1)));
         assertThat(alertsRaised.get(0).getUri(), is(equalTo("https://example0/auth")));
-        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("sanitizedtoken10")));
+        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("PHPSESSID")));
         assertThat(alertsRaised.get(0).getOtherInfo(), is(equalTo("\ncookie:PHPSESSID")));
         assertThat(alertsRaised.get(0).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
     }
@@ -304,7 +304,7 @@ class SessionDetectionScanRuleUnitTest extends PassiveScannerTest<SessionDetecti
         // Then
         assertThat(alertsRaised.size(), is(equalTo(1)));
         assertThat(alertsRaised.get(0).getUri(), is(equalTo("https://example0/login")));
-        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("sanitizedtoken1")));
+        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("session")));
         assertThat(
                 alertsRaised.get(0).getOtherInfo(),
                 is(equalTo("\ncookie:session\ncookie:AWSALBCORS\ncookie:AWSALB")));
@@ -334,19 +334,19 @@ class SessionDetectionScanRuleUnitTest extends PassiveScannerTest<SessionDetecti
         // Then
         assertThat(alertsRaised.size(), is(equalTo(3)));
         assertThat(alertsRaised.get(0).getUri(), is(equalTo("https://example0/sign_in")));
-        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("sanitizedtoken0")));
+        assertThat(alertsRaised.get(0).getEvidence(), is(equalTo("_mastodon_session")));
         assertThat(alertsRaised.get(0).getOtherInfo(), is(equalTo("\ncookie:_mastodon_session")));
         assertThat(alertsRaised.get(0).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
 
         assertThat(alertsRaised.get(1).getUri(), is(equalTo("https://example0/")));
-        assertThat(alertsRaised.get(1).getEvidence(), is(equalTo("sanitizedtoken5")));
+        assertThat(alertsRaised.get(1).getEvidence(), is(equalTo("_session_id")));
         assertThat(
                 alertsRaised.get(1).getOtherInfo(),
                 is(equalTo("\ncookie:_session_id\ncookie:_mastodon_session")));
         assertThat(alertsRaised.get(1).getConfidence(), is(equalTo(Alert.CONFIDENCE_MEDIUM)));
 
         assertThat(alertsRaised.get(2).getUri(), is(equalTo("https://example0/")));
-        assertThat(alertsRaised.get(2).getEvidence(), is(equalTo("sanitizedtoken4747")));
+        assertThat(alertsRaised.get(2).getEvidence(), is(equalTo("_session_id")));
         assertThat(
                 alertsRaised.get(2).getOtherInfo(),
                 is(equalTo("\ncookie:_session_id\ncookie:_mastodon_session")));